backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and vector operations
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pgvector==0.2.4

# Embeddings and ML
sentence-transformers==2.7.0
numpy>=1.21.0

# Data processing
pandas>=1.5.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
tqdm>=4.60.0
requests==2.31.0
toml==0.10.2  # For prompt templates (POML)
redis==5.0.1
prometheus-client==0.17.1
opentelemetry-api==1.23.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

